python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["karellen-kim <karellen.kim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aerospike = "^11.1.0"
pyspark = "^3.4.1"
pytest-benchmark = "^4.0.0"
load-dotenv = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
